terminate with clear db

+++ tests/integration/utils.js
@@ -34,10 +34,8 @@ const createUser = async (app, token) => {
   return res.body.id
 }
 
-const destroyDB = async app => {
-  if (process.env.POSTGRE_INSTANCE) {
-    app.clearPostgresDB()
-  } else if (process.env.NODE_ENV === 'test') {
+const destroyDB = async () => {
+  if (!process.env.POSTGRE_INSTANCE && process.env.NODE_ENV === 'test') {
     await fs.unlinkSync('./test.sqlite3')
   }
 }
